refactor(roles): replace deep imports with public API imports (#2441)
Replace 20 non-public API imports in src/vibe3/roles/ with public API
imports from parent modules (config, services, clients, execution).
